Move InventoryUtils from Features to Utils and update refs (khanshoaib3/stardew-access#112). Cleanup unnecessary using directives.
