+++ stardew-access/API.cs
@@ -1,5 +1,4 @@
 using Microsoft.Xna.Framework;
-using stardew_access.Features;
 using stardew_access.Utils;
 
 namespace stardew_access
