+++ stardew-access/Features/StaticTiles.cs
@@ -1,9 +1,8 @@
 using StardewValley;
 using System.Text.Json;
-using stardew_access.Utils;
 using static stardew_access.Utils.JsonLoader;
 
-namespace stardew_access.Features
+namespace stardew_access.Utils
 {
     public class StaticTiles
     {
