+++ stardew-access/Features/Warnings.cs
@@ -1,6 +1,4 @@
-using stardew_access.Utils;
-
-namespace stardew_access.Features
+namespace stardew_access.Utils
 {
     /// <summary>
     /// Warns the player when their health or stamina/energy is low. Also warns when its past midnight.
