+++ stardew-access/Patches/DonationMenuPatches/MuseumMenuPatch.cs
@@ -1,6 +1,6 @@
 using Microsoft.Xna.Framework;
 using StardewValley;
-using stardew_access.Features;
+using stardew_access.Utils;
 using StardewValley.Locations;
 using StardewValley.Menus;
