+++ stardew-access/Patches/GameMenuPatches/CraftingPagePatch.cs
@@ -1,5 +1,5 @@
 using StardewValley;
-using stardew_access.Features;
+using stardew_access.Utils;
 using StardewValley.Menus;
 using StardewValley.Objects;
