+++ stardew-access/Patches/MiscPatches/IClickableMenuPatch.cs
@@ -1,4 +1,4 @@
-using stardew_access.Features;
+using stardew_access.Utils;
 using StardewValley;
 using StardewValley.Menus;
 using System.Text;
