+++ stardew-access/Utils/JsonLoader.cs
@@ -1,4 +1,3 @@
-using System.IO;
 using System.Text.Json;
 
 namespace stardew_access.Utils
